+++ facile_rs/utils/__init__.py
@@ -1,5 +1,7 @@
 import logging
 import os
+import shutil
+import tempfile
 from pathlib import Path
 
 from dotenv import load_dotenv
@@ -71,3 +73,16 @@ class Settings:
 
 
 settings = Settings()
+
+
+def setup_assets_path(assets_path, remove_existing=False, exist_ok=False):
+    path = Path(assets_path).expanduser()
+    if remove_existing:
+        shutil.rmtree(path)
+    path.mkdir(parents=True, exist_ok=exist_ok)
+    return path
+
+
+def setup_tmp_assets_path():
+    tmp_dir = tempfile.TemporaryDirectory(prefix='facile-rs', delete=False)
+    return Path(tmp_dir.name), tmp_dir